+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@
   - overhauled command pipeline based on new and more flexible command bus middlewares
   - CommandExecutionOptions that can change the tenant context to run the command in and other options
 - AsyncEventListenerBindingExtensions for easier async event listener registration
+- convenience methods IReadRepository.GetManyAsync<T> and IReadRepository.FindManyAsync<T> using Guid IDs
 
 ### Changed
 - breaking change: null tenants now cannot access other tenant's data by default (can be changed by configuration)